+++ src/main/kotlin/dev/tailwindrainbow/intellij/adapter/intellij/highlighting/TailwindRainbowAnnotator.kt
@@ -18,7 +18,9 @@ class TailwindRainbowAnnotator : Annotator {
         element: PsiElement,
         holder: AnnotationHolder,
     ) {
-        if (element !is PsiFile) return
+        if (element !is PsiFile) {
+            return
+        }
 
         val extension = element.virtualFile?.extension ?: return